requirements.txt
flask
werkzeug
langchain
openai
python-dotenv
boto3
chromadb
pypdf
tiktoken
unstructured
langchain-community
langchain_openai
langchain_chroma